Ghunna rule
The highlighting for this rule should include the letter before ghunna occurs (previous prononced letter). This is because it cues the reader to do ghunna. The highlighting should end until the beginning of the next letter i.e. including all diacritic marks of the noon/meem involved. This takes this into account.