Refactoring in preparation for testing and export.
Each Rule is expected to now return a List<Result>, which can now either
be a Result or a TwoPartResult (depending on the number of highlights
that are part of that particular rule).

Insha'Allah this will enable us to write tests for particular rules, in
addition to writing a pluggable system for exporting the data in the
future.