Remove overlapping entries in the sorting stage
The html exporter used substrings to apply rules to the html output. If there was an overlap then the parts of the substring may repeat. No matter what the exporter, there would always be a need to handle overlaps. Overlaps happen frequently with ghunna vs idghaam vs meem idghaam see issue #11. In this workaround, overlaps are removed all together. Also, ghunna is given the least preference and is always removed. Doing this at the sorting level will help remove the need of handling this in any future exporters. .