Correctly handle "pronounce like meem" character
On top of a noon sakin that will be pronounced as a meem, the madani
mushaf and uthmani script adds a small meem above the noon (this is not
to be confused with the "stopping" meem). When looking for a noon sakina
followed by a ba' for iqlab, don't break the noon's "sakina-ness" due to
the pronunciation guide meem.