Remove buggy zero width hack
Arabic characters separated by spans are broken on webkit. While there
is a hack to make this work (inserting zero width joining characters in
certain places), the output is still not normal on WebKit when compared
to something like Firefox without the hack. This, combined with the fact
that the hack was incorrectly implemented (and thus Chrome and Firefox
were often broken or non-ideal), were the causes of removing this hack.